backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
sqlalchemy==2.0.31
psycopg[binary]==3.3.3
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic-settings==2.3.4
openai==1.40.6
openpyxl==3.1.5
boto3==1.34.149
python-docx==1.1.2
pypdf==4.3.1
reportlab==4.2.2
apscheduler==3.10.4
email-validator==2.2.0